refactor: split build-path transforms into build_transforms.py
Lift the four framework transforms whose output is baked into the cached
L2 parquet out of transformations.py (read-path) into a new
build_transforms.py, re-exporting them so the ~30 country build scripts
that import them from lsms_library.transformations keep working unchanged.

Moved (byte-identical bodies, source order preserved so the registry's
dependency on fill_v_with_coord_bin holds):
  - food_acquired_to_canonical
  - _finalize_canonical_food_acquired
  - fill_v_with_coord_bin
  - apply_derived (+ _DERIVED_TRANSFORMERS registry)

Pure structural change, no behavior change: both import paths resolve to
the same objects; build_transforms.py depends only on pandas/numpy, so no
circular import. validate_acquisition_source stays in transformations.py —
prototyping showed it has zero callers (its docstring claims a
_finalize_result call that does not exist), so it is a dormant
read-path-intended validator, not a build-path transform.

Step 1 of the build/read split discussed on PR #531: make the boundary
structural so the v0.8.0 content-hash cache layer can version build-path
transform code precisely. Step 2 (fold build_transforms.py's source hash
into _table_cache_hash / _input_hash, demote LSMS_CACHE_SCHEMA) is a
separate follow-up.
